Inline ServerStateImpl into ProtegeServer
This parent class serves no purpose. Its implementation isn't used anywhere
apart from the subclass and any abstraction boundary it might provide is already
provided by the interface. Hence it's implementation is moved to the only place
where it's used - the subclass.
